requirements.txt
llama-cpp-python
PyQt6